Renamed StringAnonymizer to FixedValueAnonymizer
