+++ Anonymizer/FixedValueAnonymizer.php
@@ -12,26 +12,60 @@
 
 namespace SuperBrave\GdprBundle\Anonymizer;
 
+use SuperBrave\GdprBundle\Manipulator\PropertyManipulator;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
+
 /**
- * Class FixedValueAnonymizer
+ * Email anonymizer class
  *
  * @package SuperBrave\GdprBundle\Anonymizer
  */
 class FixedValueAnonymizer implements AnonymizerInterface
 {
+    /**
+     * Property Manipulator service
+     *
+     * @var PropertyManipulator
+     */
+    private $propertyManipulator;
+
+    /**
+     * StringAnonymizer constructor.
+     * @param PropertyManipulator $propertyManipulator
+     */
+    public function __construct(PropertyManipulator $propertyManipulator)
+    {
+        $this->propertyManipulator = $propertyManipulator;
+    }
+
+    /**
+     * @param mixed $propertyValue The value that has to be converted
+     * @param array $options Options to help the anonymizer do its job
+     *
+     * @return string               The Anonymized string
+     */
     public function anonymize($propertyValue, array $options = [])
     {
         $resolver = new OptionsResolver();
         $this->configureOptions($resolver);
         $options = $resolver->resolve($options);
 
-        return $options['annotationValue'];
+        $annotationValue = $options['annotationValue'];
+
+        $matches = [];
+        if (preg_match_all('/{(.*?)}/', $options['annotationValue'], $matches)) {
+            foreach ((array)$matches[1] as $key => $property) {
+                $wildcardValue = $this->propertyManipulator->getPropertyValue($options['object'], $property);
+                $annotationValue = str_replace($matches[0][$key], $wildcardValue, $annotationValue);
+            }
+        }
+
+        return $annotationValue;
     }
 
     private function configureOptions(OptionsResolver $resolver)
     {
-        $resolver->setRequired(['annotationValue']);
+        $resolver->setRequired(['annotationValue', 'object']);
     }
 }
